DeliveryCivil/requirements.txt
# Python ETL Pipeline - DeliveryCivil
# Versão mínima: Python 3.11+

# Core ETL
pandas>=2.0.0
polars>=0.19.0  # Opcional: alternativa rápida ao pandas
numpy>=1.24.0

# Validação e Modelos
pydantic>=2.0.0
pydantic-settings>=2.0.0
great-expectations>=0.18.0  # Opcional: validação avançada de qualidade

# Requisições HTTP
requests>=2.31.0
httpx>=0.24.0  # Opcional: cliente HTTP assíncrono

# Bancos de Dados
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL (se necessário)
pymongo>=4.5.0  # MongoDB (se necessário)
pyodbc>=4.0.39  # SQL Server

# Arquivos
openpyxl>=3.1.0  # Excel
pyarrow>=12.0.0  # Parquet
python-dateutil>=2.8.0

# Power BI Integration - Removido (usando iframe público agora)

# Orquestração (escolher uma opção)
# Opção 1: Apache Airflow
apache-airflow>=2.7.0

# Opção 2: Dagster (alternativa moderna)
# dagster>=1.5.0
# dagster-webserver>=1.5.0

# Opção 3: Prefect (alternativa simples)
# prefect>=2.14.0

# API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6  # Para upload de arquivos

# Front-end (escolher uma opção)
# Opção 1: Streamlit
streamlit>=1.28.0
streamlit-aggrid>=0.3.4

# Opção 2: Dash com Bootstrap
dash>=2.14.0
dash-bootstrap-components>=1.5.0
plotly>=5.17.0

# Utilitários
python-dotenv>=1.0.0  # Gerenciamento de variáveis de ambiente
loguru>=0.7.0  # Logging avançado (opcional, alternativa ao logging padrão)

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

